+++ app/models/conversion/project.rb
@@ -15,4 +15,19 @@ class Conversion::Project < Project
 
     conversion_dates.order(:created_at).first.previous_date
   end
+
+  def all_conditions_met?
+    tasks_data.conditions_met_confirm_all_conditions_met?
+  end
+
+  def conversion_date_confirmed_and_passed?
+    !conversion_date_provisional? && conversion_date.past?
+  end
+
+  def grant_payment_certificate_received?
+    user = assigned_to
+    tasks = Conversion::Task::ReceiveGrantPaymentCertificateTaskForm.new(tasks_data, user)
+    return true if tasks.status.eql?(:completed)
+    false
+  end
 end