+++ spec/models/conversion/project_spec.rb
@@ -54,4 +54,99 @@ RSpec.describe Conversion::Project do
       end
     end
   end
+
+  describe "#all_conditions_met?" do
+    context "when the all conditions met task is completed" do
+      let(:tasks_data) { create(:conversion_tasks_data, conditions_met_confirm_all_conditions_met: true) }
+      let(:project) { build(:conversion_project, tasks_data: tasks_data) }
+
+      it "returns true" do
+        expect(project.all_conditions_met?).to eq(true)
+      end
+    end
+
+    context "when the all conditions met task has not been completed" do
+      let(:tasks_data) { create(:conversion_tasks_data, conditions_met_confirm_all_conditions_met: nil) }
+      let(:project) { build(:conversion_project, tasks_data: tasks_data) }
+
+      it "returns false" do
+        expect(project.all_conditions_met?).to eq(false)
+      end
+    end
+  end
+
+  describe "#opened?" do
+    let(:project) {
+      build(:conversion_project,
+        conversion_date: conversion_date,
+        conversion_date_provisional: conversion_date_provisional)
+    }
+
+    context "when the conversion date has not passed" do
+      let(:conversion_date) { Date.tomorrow }
+      let(:conversion_date_provisional) { false }
+
+      it "returns false" do
+        expect(project.conversion_date_confirmed_and_passed?).to be false
+      end
+    end
+
+    context "when the conversion date has passed but the conversion date is provisional" do
+      let(:conversion_date) { Date.yesterday }
+      let(:conversion_date_provisional) { true }
+
+      it "returns false" do
+        expect(project.conversion_date_confirmed_and_passed?).to be false
+      end
+    end
+
+    context "when the conversion date is provisional" do
+      let(:conversion_date) { Date.tomorrow }
+      let(:conversion_date_provisional) { true }
+
+      it "returns false" do
+        expect(project.conversion_date_confirmed_and_passed?).to be false
+      end
+    end
+
+    context "when the conversion date is confirmed and the conversion date has passed" do
+      let(:conversion_date) { Date.yesterday }
+      let(:conversion_date_provisional) { false }
+
+      it "returns true" do
+        expect(project.conversion_date_confirmed_and_passed?).to be true
+      end
+    end
+  end
+
+  describe "#grant_payment_certificate_received?" do
+    let(:user) { create(:user) }
+    let(:project) { build(:conversion_project, tasks_data: tasks_data) }
+
+    context "when the ReceiveGrantPaymentCertificateTaskForm is NOT completed" do
+      let(:tasks_data) {
+        create(:conversion_tasks_data,
+          receive_grant_payment_certificate_check_and_save: nil,
+          receive_grant_payment_certificate_update_kim: nil,
+          receive_grant_payment_certificate_update_sheet: nil)
+      }
+
+      it "returns false" do
+        expect(project.grant_payment_certificate_received?).to be false
+      end
+    end
+
+    context "when the ReceiveGrantPaymentCertificateTaskForm is completed" do
+      let(:tasks_data) {
+        create(:conversion_tasks_data,
+          receive_grant_payment_certificate_check_and_save: true,
+          receive_grant_payment_certificate_update_kim: true,
+          receive_grant_payment_certificate_update_sheet: true)
+      }
+
+      it "returns true" do
+        expect(project.grant_payment_certificate_received?).to be true
+      end
+    end
+  end
 end