+++ spec/models/project_spec.rb
@@ -336,26 +336,6 @@ RSpec.describe Project, type: :model do
     end
   end
 
-  describe "#all_conditions_met?" do
-    context "when the all conditions met task is completed" do
-      let(:tasks_data) { create(:conversion_tasks_data, conditions_met_confirm_all_conditions_met: true) }
-      let(:project) { build(:conversion_project, tasks_data: tasks_data) }
-
-      it "returns true" do
-        expect(project.all_conditions_met?).to eq(true)
-      end
-    end
-
-    context "when the all conditions met task has not been completed" do
-      let(:tasks_data) { create(:conversion_tasks_data, conditions_met_confirm_all_conditions_met: nil) }
-      let(:project) { build(:conversion_project, tasks_data: tasks_data) }
-
-      it "returns false" do
-        expect(project.all_conditions_met?).to eq(false)
-      end
-    end
-  end
-
   describe "Scopes" do
     describe "conversions" do
       before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }