+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Do not allow a project to be completed until all conditions have been met, the
+  grant certificate is returned and the academy has opened
+
 ### Changed
 
 ### Fixed